# backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: PVA Bazaar Marketplace API
  description: Complete global marketplace for artisans, creators, and sellers with all 8 phases
  version: 8.0.0
  contact:
    name: PVA Support
    url: https://pvabazaar.org
  license:
    name: Proprietary
    url: https://pvabazaar.org/license
servers:
  - url: https://api.pvabazaar.org
    description: Production
  - url: http://localhost:5001
    description: Development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: http
      scheme: bearer
      description: API Key for partner integrations
...
security:
  - bearerAuth: []
paths:
...
  /api/messages:
    post:
      tags: [Messaging - Phase 4]
      summary: Send message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipientId: { type: string }
                message: { type: string }
      responses:
        '201':
          description: Message sent

    get:
      tags: [Messaging - Phase 4]
      summary: Get conversations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Conversations list
...